Fix mistakes + convert to 16 bits integers

+++ src/daemon.h
@@ -15,7 +15,7 @@
 
 typedef struct _meteodata {
 	struct timeval time;
-	float data;
+	int16_t data;
 } MeteoData;
 
 enum { TEMP, LIGHT };
